Make App Fee float

+++ config/initializers/cake_coupon_book.rb
@@ -1,4 +1,4 @@
 module CakeCouponBook
-  APPLICATION_FEE = ENV['APPLICATION_FEE'].to_i
+  APPLICATION_FEE = ENV['APPLICATION_FEE'].to_f
   STRIPE_FEE = 2.9
 end
